scripts/importador/requirements.txt
fdb==1.8.1
requests==2.31.0
python-dotenv==1.0.1
tqdm==4.66.2
